refactor contract utilities into separate mod
